requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-learn>=0.24.0
joblib>=1.0.0
pandas>=2.0.0
Flask>=2.3.0
python-multipart>=0.0.6
python-dotenv>=0.21.0
Werkzeug>=2.3.0
pdf2image>=1.16.0
pytesseract>=0.3.10
openai>=1.0.0
insightface>=0.7.0
onnxruntime>=1.15.0
flask-cors>=4.0.0
Flask-WTF>=1.2.0
Flask-Limiter>=3.5.0
psycopg2-binary>=2.9.9
azure-storage-blob>=12.19.0
